chore: move card to component
